+++ src/golang/index.ts
@@ -1,2 +1,3 @@
 export * as strconv from './strconv/index.ts'
 export * as strings from './strings/index.ts'
+export * as time from './time/index.ts'